Fix incorrect Header import
- Adjust Header component to not include any of SearchBar
- Re-create SearchBar component from Header
